finished authorization

+++ backend/routes/classes.js
@@ -33,7 +33,8 @@ router.post("", checkAuth, multer({storage: storage}).single("image"), (req, res
     className: req.body.className,
     classWeight: req.body.classWeight,
     classDes: req.body.classDes,
-    imagePath: url + "/images/" + req.file.filename
+    imagePath: url + "/images/" + req.file.filename,
+    creator: req.userData.userId //userid can be decoded from the token
   });
   // save() saves into mongodb
   addedClass.save().then(createdClass => {
@@ -62,11 +63,16 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     className: req.body.className,
     classWeight: req.body.classWeight,
     classDes: req.body.classDes,
-    imagePath: imagePath
+    imagePath: imagePath,
+    creator: req.userData.userId
   });
-  Class.updateOne({_id: req.params.id}, updatedClass).then(result => {
+  Class.updateOne({_id: req.params.id, creator: req.userData.userId }, updatedClass).then(result => {
     console.log(result);
-    res.status(200).json({message: "Post updated brah"});
+    if (result.nModified > 0){
+      res.status(200).json({message: "Post updated brah"});
+    } else {
+      res.status(401).json({message: "Ya not authorized edit brah"});
+    }
   });
 });
 
@@ -103,9 +109,13 @@ router.get("/:id", (req, res, next) => {
   })
 })
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Class.deleteOne({_id: req.params.id}).then(result => {
-    console.log(result);
-    res.status(200).json({message: "Post deleted brah"});
+  Class.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+    // console.log("result is: ", result);
+    if (result.deletedCount > 0){
+      res.status(200).json({message: "Post deleted brah"});
+    } else {
+      res.status(401).json({message: "Ya not authorized to delete brah" + result.n});
+    }
   });
 });
